Re-worked a few bits of initialisation to reduce memory usage to when its needed instead of having static initialisation when these plug-ins aren't likely going to be actively used in most wacup installs Also re-packed some of the plug-in structures to avoid memory holes more so for the x64 based builds (hence the double to float volume change being the most obvious of the changes made related to this)

+++ src/in_gsf/vbam/apu/Blip_Buffer.cpp
@@ -215,13 +215,13 @@ void blip_eq_t::generate(float *out, int count) const
 {
 	// lower cutoff freq for narrow kernels with their wider transition band
 	// (8 points->1.49, 16 points->1.15)
-	double oversample = blip_res * 2.25 / count + 0.85;
+	double oversample = (double)blip_res * 2.25 / count + 0.85;
 	double half_rate = this->sample_rate * 0.5;
 	if (this->cutoff_freq)
 		oversample = half_rate / this->cutoff_freq;
 	double cutoff = this->rolloff_freq * oversample / half_rate;
 
-	gen_sinc(out, count, blip_res * oversample, this->treble, cutoff);
+	gen_sinc(out, count, (double)blip_res * oversample, this->treble, cutoff);
 
 	// apply (half of) hamming window
 	double to_fraction = M_PI / (count - 1);
